Fix type for SetpointCommandScaled
Was c_float, should be c_int
Added test for SetpointCommandScaled

+++ lib60870-Python/lib60870/information_object.py
@@ -1679,7 +1679,7 @@ class SetpointCommandScaled(ctypes.Structure, IOBase):
         return lib.SetpointCommandScaled_create(
             pSetpointCommandScaled(self),
             c_int(ioa),
-            c_float(value),
+            c_int(value),
             c_bool(selectCommand),
             c_int(ql)).contents
